feat: Add "Agendas e Cadernos" category and special offer
Introduces a new product category "AGENDAS_CADERNOS" to the system, updating type definitions and the enum for categories.

Additionally, a special offer product for "Lápis de Cor Faber-Castell" has been added to constants and will be presented to the user in the cart if not already present. Cart drawer logic is updated to handle the special offer presentation.